+++ skeleton_files/skeleton.py
@@ -1,5 +1,5 @@
 # class Pot:
-#     value: float
+#     value: int
 #     players: list[str]
 
 
@@ -8,12 +8,12 @@
 #     index_of_small_blind: int
 #     players: list[str]
 #     player_cards: list[str]
-#     held_money: list[float]
-#     bet_money: list[float]
+#     held_money: list[int]
+#     bet_money: list[int]
 #     community_cards: list[str]
 #     pots: list[Pot]
-#     small_blind: float
-#     big_blind: float
+#     small_blind: int
+#     big_blind: int
 
 # ======================= #
 # ACTUAL BOT CODE HERE!!! #
@@ -26,8 +26,8 @@ def set_state_input(state: GameState):
     state.index_of_small_blind = int(input())
     state.players = input().split()
     state.player_cards = input().split()
-    state.held_money = list(map(float, input().split()))
-    state.bet_money = list(map(float, input().split()))
+    state.held_money = list(map(int, input().split()))
+    state.bet_money = list(map(int, input().split()))
     state.community_cards = input().split()
 
     num_pots = int(input())
@@ -35,12 +35,12 @@ def set_state_input(state: GameState):
     for _ in range(num_pots):
         pot_input_list = input().split()
         p = Pot()
-        p.value = float(pot_input_list[0])
+        p.value = int(pot_input_list[0])
         p.players = pot_input_list[1:]
         state.pots.append(p)
 
-    state.small_blind = float(input())
-    state.big_blind = float(input())
+    state.small_blind = int(input())
+    state.big_blind = int(input())
 
 
 def main():
