+++ src/admin.py
@@ -97,7 +97,7 @@ async def make_move_on_tables(
     description="for human input",
 )
 async def make_move_on_table(
-    table_id: str, raise_size: float, _: User = Depends(verify_admin_user)
+    table_id: str, raise_size: int, _: User = Depends(verify_admin_user)
 ):
     try:
         return await tournament.make_moves([table_id], [raise_size])
